Base Settings Changes
Added Zappa settings to base settings file for working with a custom domain name and a custom cloudfront distribution. The solution was recommended in the Github issue here: https://github.com/Miserlou/Zappa/issues/1084

These settings tell django to use the X-Forwarded-Host header which we can add as a custom header in the cloudfront distribution.

Script names also also forced to an empty string to strip out the stage name of the zappa environment. This is the only method I know of and could included potential side-effects so something to be aware of.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>