EventIndexPage Cloudfront Cache Invalidation
Added methods to EventIndexPage to aid in the collection of urls that may need their caches invalidated after a page is revised.

Added signal handling functions in handlers.py to match EventIndexPages that contain a revised page and need their cache invalidated as well.

Registered signal in apps.py and made django aware of this config in the events module __init__.py file.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>