Fixed IDing of Block Template For Loops
Button list and image list block templates formerly used only the loop counter as an ID, causing a bug where multiple blocks elements would share the same ID on a page.

To solve this, changed all the page templates using those blocks to pass the block id to the block templates. This ID was then added to the html element id in additional to the for loop counter, so all elements would have a unique ID.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>