Added new EventGeneralPage Model
Added new model to events app that inherits from the GeneralPage of the home app. The only difference between this new model it's parent is that it adds the events_dates and notes fields, similar to the EventPage model.

EventGeneralPage.image is added as an alias of GeneralPage.thumbnail so that the template of EventIndexPage doesn't need to be changed to handle EventGeneralPage children.

EventIndexPage get_event_items method no longer specifies object type since it can vary between EventPage and EventGeneralPage.

Registered EventGeneralPage with the cache invalidator handler.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>