Banner Image Style Change
Removed "responsive" class from banner images because it created problems when small images are used.

Instead, specified max dimensions of 1280x500 which works well for both images natively sized to banner dimensions and images with larger heights.

To emulate the centering that the responsive class provided, the text-center class was added to the parent div of the banner element.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>