Policy Page Snippet
Since so many event pages include a link to the refund and cancellation policy, it makes sense to create snippets for these links. The new snippet can also be used to link to other similar type policy pages.

The new PolicyLink snippet includes fields for the name of the snippet, the text to be displayed, and a page chooser which the displayed text will link to.

The snippet is now included in the EventPage model. The template for this model has been updated to the snippet below the body content of the page.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>