Concert Price Changes
Concert member and public price properties changed from IntegerFields to CharFields to facilitate price ranges and generally allow for greater flexibility in how prices are displayed.

Member price was changed to become optional for situations when there is no difference in cost between members and the public. The template was updated to reflect this, such that when member price is null, only the price is displayed rather than including text distinguishing between member/public prices.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>