Virtual Concert Date Handling
Added a field to record the date that virtual concerts will remain available until.

Added two new properties to the Concert model for comparing the current date with the available_until field and the concert_date field.

Modified template to make use of date comparator properties. Concerts that occurred in the past include a CSS class to lower their opacity. For virtual concerts whose live performance occurred in the past, the concert date and times are replaced by text indicating how long the concert will be available on-demand.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>