EventPage Refactoring
To account for the fact that EventIndexPages can now appear under other EventIndexPages, it made sense to change EventPage to also inherit from AbstractBase and transition away from the use of the image property to instead use the thumbnail property.

To ease the transition for existing pages, a save override was created that sets thumbnail equal to image if it is not already set.

Page templates were updated to expect only the thumbnail property.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>