Change to Serializer Validation and Collection handling
Collection serializer does not enforce validation of unique plant_id to make it possible to POST the same collection with updated records.

Collection creation searches for an existing collection by plant_id. If found, it will update that record with the provided data or create a new record if it doesn't exist.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>